backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.6
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langsmith==0.1.63
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.30.4
orjson==3.10.3
packaging==23.2
parse==1.20.2
pinecone-client==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
sniffio==1.3.1
soupsieve==2.5
sql==2022.4.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
